+++ municipalityinitiative-web/src/test/java/fi/om/municipalityinitiative/service/UserServiceIntegrationTest.java
@@ -79,7 +79,7 @@ public class UserServiceIntegrationTest extends ServiceIntegrationTestBase{
         userService.authorLogin(testHelper.getPreviousTestManagementHash(), requestMock);
         LoginUserHolder loginUserHolder = userService.getRequiredLoginUserHolder(requestMock);
 
-        assertThat(loginUserHolder.getNormalLoginUser().getAuthorId().toLong(), is(testHelper.getLastNormalAuthorId()));
+        assertThat(loginUserHolder.getNormalLoginUser().getAuthorId(), is(testHelper.getLastNormalAuthorId()));
         assertThat(loginUserHolder.getUser().hasRightToInitiative(initiative), is(true));
         assertThat(loginUserHolder.getUser().hasRightToInitiative(-1L), is(false));
     }
